- Beginning of major changes to open/closed loop stimulation functionality - the .olstim format has changed - makestimfile.m is changed to produce a more efficient .olstim - stimulation time base is now i 1/10000ths of a second - stimbuffer is shedding its load all the data at once stuff in favor of Riley's append methods - stimbuffer needs events telling other processes when it is finished loading all the required stimulus to the NI Buffer and when it has finished loading X stimuli -changes to the GUI that are not yet functional
git-svn-id: https://neurorighter.googlecode.com/svn/branches/NeuroRighter@232 471701b0-e017-11dd-9759-139c9ab14fc6

+++ NeuroRighter_05/Stimulation/ClosedLoopUtils.cs
@@ -105,7 +105,7 @@ namespace NeuroRighter
            
         }
 
-        public void appendStim(double[] timeVec, int[] channelVec, double[,] waveMat)
+        public void appendStim(int[] timeVec, int[] channelVec, double[,] waveMat)
         {
             buffer.append(timeVec, channelVec, waveMat);
         }
